requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.11.0
appdirs==1.4.4
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bce-python-sdk==0.9.46
beautifulsoup4==4.13.5
Brotli==1.1.0
cachetools==6.2.1
camel==0.1.2
camel-ai==0.2.78
cbor==1.0.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
datasets==4.3.0
decorator==5.2.1
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docling==2.50.0
docling-core==2.47.0
docling-ibm-models==3.9.1
docling-parse==4.4.0
docstring_parser==0.17.0
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Faker==37.6.0
filelock==3.19.1
filetype==1.2.0
FlagEmbedding==1.3.5
flatbuffers==25.2.10
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
google-auth==2.41.1
google-genai==1.46.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
html2image==2.0.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.4
humanfriendly==10.0
identify==2.6.15
idna==3.10
ijson==3.4.0.post0
imageio==2.37.0
imagesize==1.4.1
imgkit==1.2.3
importlib_metadata==8.7.0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
latex2mathml==3.78.1
lazy_loader==0.4
lightning-utilities==0.15.2
lxml==5.4.0
lz4==4.4.4
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==0.13.1
marker-pdf==1.7.1
marko==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mcp==1.19.0
mdurl==0.1.2
modelscope==1.31.0
more-itertools==10.8.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime @ file:///mnt/gaojuanru/Paper2Poster/onnxruntime-1.22.1-cp310-cp310-manylinux1_x86_64.manylinux2014_x86_64.whl#sha256=bddc75868bcf6f9ed76858a632f65f7b1846bdcefc6d637b1e359c2c68609964
open_clip_torch==3.2.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0
paddleocr==3.2.0
paddlepaddle-gpu==2.6.1
paddlex==3.2.1
pandas==2.3.2
pdftext==0.6.3
peft==0.17.1
pillow==10.4.0
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
plum-dispatch==1.7.4
polyfactory==2.22.2
pre_commit==4.3.0
prettytable==3.16.0
propcache==0.4.1
protobuf==6.32.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
pyee==11.1.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.2.5
pypdfium2==4.30.0
pyphen==0.17.2
pyppeteer==2.0.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
RapidFuzz==3.14.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree @ file:///home/conda/feedstock_root/build_artifacts/rtree_1755128329929/work
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
selenium==4.37.0
semchunk==2.2.2
sentence-transformers==5.1.1
sentencepiece @ file:///home/conda/feedstock_root/build_artifacts/sentencepiece-split_1758536863393/work/python
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spire-doc==13.3.7
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.7.0
timm==1.0.21
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
torch==2.6.0
torch-fidelity==0.3.0
torchmetrics==1.8.2
torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.1
trec-car-tools==2.6
trio==0.31.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
unlzw3==0.2.3
urllib3==1.26.20
uvicorn==0.38.0
virtualenv==20.35.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.14
weasyprint==66.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==10.4
wsproto==1.2.0
xlsxwriter==3.2.5
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zlib-state==0.1.10
zopfli==0.2.3.post1
